pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.8",
    "django-crispy-forms>=2.5",
    "djangorestframework>=3.16.1",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.11",
]
